Move public func to conftest, add room testcase.
